gitypeinfo: Use gssize to return the array length index
Use a bigger integer value, even though int is more than enough, but
to make it clearer that we're returning a size-related value.

+++ girepository/gitypeinfo.h
@@ -96,7 +96,7 @@ GI_AVAILABLE_IN_ALL
 GIBaseInfo *           gi_type_info_get_interface       (GITypeInfo *info);
 
 GI_AVAILABLE_IN_ALL
-int                    gi_type_info_get_array_length_index (GITypeInfo *info);
+gssize                 gi_type_info_get_array_length_index (GITypeInfo *info);
 
 GI_AVAILABLE_IN_ALL
 gssize                 gi_type_info_get_array_fixed_size (GITypeInfo *info);
